requirements.txt
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.2
